リファクタ_ApplicationControllerに関するファイルをメンテナンス
